Number Guessing Game
A user can input a number. This input should be compared with a set of numbers in a random variable (for example, between 1 and 100. You can set any limit you want). If it matches, it should indicate the Guessed number is correct and if not, another input should be taken.this should be taken until the Guessed number is correct.